+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/adapter/AnnotationHandlerAdapter.java
@@ -1,18 +1,19 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc.tobe.adapter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import com.interface21.webmvc.servlet.ModelAndView;
+import com.interface21.webmvc.servlet.mvc.tobe.handler.HandlerExecution;
 
-public class ErrorHandlerAdapter implements HandlerAdapter {
+public class AnnotationHandlerAdapter implements HandlerAdapter {
 
     @Override
     public boolean supports(Object handler) {
-        return handler instanceof ErrorController;
+        return handler instanceof HandlerExecution;
     }
 
     @Override
     public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        return ((ErrorController) handler).notFoundErrorHandle(request, response);
+        return ((HandlerExecution) handler).handle(request, response);
     }
 }
