+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/adapter/HandlerAdapterRegistry.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc.tobe.adapter;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,12 +6,6 @@ import java.util.List;
 public class HandlerAdapterRegistry {
 
     private final List<HandlerAdapter> handlerAdapters = new ArrayList<>();
-
-    public HandlerAdapterRegistry() {
-        AnnotationHandlerAdapter annotationHandlerAdapter = new AnnotationHandlerAdapter();
-        ErrorHandlerAdapter errorHandlerAdapter = new ErrorHandlerAdapter();
-        this.handlerAdapters.addAll(List.of(annotationHandlerAdapter, errorHandlerAdapter));
-    }
 
     public void addHandlerAdapter(HandlerAdapter handlerAdapter) {
         handlerAdapters.add(handlerAdapter);
