+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/handler/HandlerKey.java
@@ -1,8 +1,7 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
-
-import com.interface21.web.bind.annotation.RequestMethod;
+package com.interface21.webmvc.servlet.mvc.tobe.handler;
 
 import java.util.Objects;
+import com.interface21.web.bind.annotation.RequestMethod;
 
 public class HandlerKey {
 
@@ -24,8 +23,10 @@ public class HandlerKey {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof HandlerKey)) return false;
+        if (this == o)
+            return true;
+        if (!(o instanceof HandlerKey))
+            return false;
         HandlerKey that = (HandlerKey) o;
         return Objects.equals(url, that.url) && requestMethod == that.requestMethod;
     }
